chore: restructure and split generators to improve code maintainability
