+++ .yo-rc.json
@@ -1,7 +1,5 @@
 {
   "generator-react-library": {
-    "promptValues": {
-      "cool": true
-    }
+    "promptValues": {}
   }
 }
