+++ generators/app/resources/constants.js
@@ -1,17 +1,7 @@
 const YARN_CHOICE = 'yarn';
 const NPM_CHOICE = 'npm';
 
-/* TODO: implement reading folder way to provide them */
-const ROOT_FILES = [
-  '.editorconfig',
-  '.eslintrc.js',
-  '.gitignore',
-  'rollup.config.js',
-  'babel.config.js',
-];
-
 module.exports = {
   NPM_CHOICE,
   YARN_CHOICE,
-  ROOT_FILES,
 };
