+++ generators/scaffold/resources/constants.js
@@ -1,0 +1,12 @@
+/* TODO: implement reading folder way to provide them */
+const ROOT_FILES = [
+  '.editorconfig',
+  '.eslintrc.js',
+  '.gitignore',
+  'rollup.config.js',
+  'babel.config.js',
+];
+
+module.exports = {
+  ROOT_FILES,
+};
