+++ generators/scaffold/templates/.editorconfig
