+++ generators/scaffold/templates/.eslintrc.js
