+++ generators/scaffold/templates/.gitignore
