+++ generators/scaffold/templates/babel.config.js
