+++ generators/scaffold/templates/rollup.config.js
