+++ generators/scaffold/templates/src/Button/index.js
