+++ generators/scaffold/templates/src/index.js
