+++ generators/testing/resources/constants.js
@@ -1,0 +1,7 @@
+const JEST_ENZYME = 'enzyme';
+const TESTING_LIBRARY = 'react-testing-library';
+
+module.exports = {
+  JEST_ENZYME,
+  TESTING_LIBRARY,
+};
