+++ generators/testing/templates/babel.config.js
