+++ generators/testing/templates/jest.config.js
