+++ package.json
@@ -19,6 +19,7 @@
     "test:w": "yarn test --watch"
   },
   "dependencies": {
+    "lodash": "^4.17.11",
     "yeoman-generator": "^4.0.1"
   },
   "devDependencies": {
